Added inheritance support to the InnerBuilder implementation
