+++ jsonschema2pojo-core/src/test/resources/schema/child.json
@@ -1,0 +1,15 @@
+{
+  "$id": "https://example.com/child.schema.json",
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "description" : "A schema definition for an object which can be inherited",
+  "javaType" : "com.example.package.GeneratedChildType",
+  "type" : "object",
+  "extends" : {
+    "$ref" : "parent.json"
+  },
+  "properties" : {
+    "childProperty" : { "type" : "string" }
+  },
+  "dependencies" : {
+  }
+}
